Refactor codebase into a python package (#32)
Make it possible to `neural-lam` as a package, thereby enabling the possibility to run from anywhere once the
package has been installed. This means it is now possible (in theory) to train neural-lam on a `.npy`-file based dataset with the neural-lam package installed into a user's `site-packages` (i.e. in their virtualenv).

The primary changes are:

- move all `*.py` that are currently outside of neural_lam/ into that folder, but keep the files the same
- change all examples of running the neural-lam "scripts", e.g. `python create_mesh.py` by `python -m neural_lam.create_mesh` in the README
- change all absolute imports to package-relative imports, i.e. `from .import utils` rather than `from neural_lam import utils`
- add tests that all the CLI entrypoints to neural_lam can be imported and add ci/cd action to run these tests

---------

Co-authored-by: SimonKamuk <43374850+SimonKamuk@users.noreply.github.com>
Co-authored-by: joeloskarsson <joel.oskarsson@liu.se>
Co-authored-by: Leif Denby <lcd@dmi.dk>