Release version 0.2.0 (#79)
## Describe your changes

Going by the [roadmap](https://github.com/mllam/neural-lam/wiki/Roadmap)
we are ready to release version 0.2.0 :partying_face:

This PR just updates the changelog for this release (and also re-orders
one entry that was put in the wrong place). Now this does the changelog
update first, and then we create the release for the commit with the
update, but I guess the order of this does not matter much.

Something I am unsure about is if we want to now remove this version
with the completed items from the roadmap? Could be good to have
archived somewhere. Perhaps we could add a section for finished releases
at the bottom.

## Github TODOs after meged:

- [ ] Create release (there is a draft release ready)
- [ ] Update roadmap

## Type of change

- [ ] 🐛 Bug fix (non-breaking change that fixes an issue)
- [ ] ✨ New feature (non-breaking change that adds functionality)
- [ ] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] 📖 Documentation (Addition or improvements to documentation)

## Checklist before requesting a review

- [x] My branch is up-to-date with the target branch - if not update
your fork with the changes from the target branch (use `pull` with
`--rebase` option if possible).
- [ ] I have performed a self-review of my code
- [ ] For any new/modified functions/classes I have added docstrings
that clearly describe its purpose, expected inputs and returned values
- [ ] I have placed in-line comments to clarify the intent of any
hard-to-understand passages of my code
- [ ] I have updated the [README](README.MD) to cover introduced code
changes
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have given the PR a name that clearly describes the change,
written in imperative form
([context](https://www.gitkraken.com/learn/git/best-practices/git-commit-message#using-imperative-verb-form)).
- [x] I have requested a reviewer and an assignee (assignee is
responsible for merging). This applies only if you have write access to
the repo, otherwise feel free to tag a maintainer to add a reviewer and
assignee.

## Checklist for reviewers

Each PR comes with its own improvements and flaws. The reviewer should
check the following:
- [ ] the code is readable
- [ ] the code is well tested
- [ ] the code is documented (including return types and parameters)
- [ ] the code is easy to maintain

## Author checklist after completed review

- [ ] I have added a line to the CHANGELOG describing this change, in a
section
  reflecting type of change (add section where missing):
  - *added*: when you have added new functionality
  - *changed*: when default behaviour of the code has been changed
  - *fixes*: when your contribution fixes a bug

## Checklist for assignee

- [x] PR is up to date with the base branch
- [x] the tests pass
- [ ] author has added an entry to the changelog (and designated the
change as *added*, *changed* or *fixed*)
- Once the PR is ready to be merged, squash commits and merge the PR.